requirements.txt
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
python-dotenv>=0.19.0
python-dotenv>=0.19.0
requests>=2.28.0